FRONT END/requirements.txt
#Python 3.9.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.3
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mysql-connector==2.2.9
numpy==1.26.4
scikit-learn==1.6.1
scipy==1.13.1
threadpoolctl==3.6.0
Werkzeug==3.1.6
zipp==3.23.0
